requirements.txt
altair==5.0.1
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
Cython==3.0.0
easyocr==1.7.0
filelock==3.9.0
fonttools==4.42.1
gitdb==4.0.10
GitPython==3.1.32
idna==3.4
imageio==2.31.2
importlib-metadata==6.8.0
importlib-resources==6.0.1
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
lapx==0.5.4
lazy_loader==0.3
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.7.2
mdurl==0.1.2
mpmath==1.2.1
networkx==3.0
numpy==1.24.1
opencv-python==4.8.0.76
packaging==23.1
pandas==2.0.3
Pillow==9.3.0
protobuf==4.24.2
psutil==5.9.5
py-cpuinfo==9.0.0
pyarrow==13.0.0
pyclipper==1.3.0.post4
pydeck==0.8.0
Pygments==2.16.1
Pympler==1.0.1
pyparsing==3.0.9
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0.1
referencing==0.30.2
requests==2.28.1
rich==13.5.2
rpds-py==0.9.2
scikit-image==0.21.0
scipy==1.11.2
seaborn==0.12.2
shapely==2.0.1
six==1.16.0
smmap==5.0.0
streamlit==1.26.0
sympy==1.11.1
tenacity==8.2.3
tifffile==2023.8.25
toml==0.10.2
toolz==0.12.0
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
tornado==6.3.3
tqdm==4.66.1
typing_extensions==4.4.0
tzdata==2023.3
tzlocal==4.3.1
ultralytics==8.0.162
urllib3==1.26.13
validators==0.21.2
watchdog==3.0.0
zipp==3.16.2
